Fixed issue with avatar on home page
On any screen with a relatively low width (~768px or less), which
includes mobile devices, my portrait would be resized to take up the
entire screen. Since the pic from gravatar is only 167px, it looked
awful, not to mention the fact that you could only see my face.

Fixed by adding a CSS rule for the already existing avatar class to
specify a max-width.