Improvements to blog template, script. Re #2
The mkblog.py script now moves any entries that aren't filed into the
<year>/<month> directory it should be in. This makes it easier to create
new entries without having to worry about placing it in the correct
folder.

**NOTE**: Even though new entries (files with the .md extension) will be
moved, any supporting files (images, etc.) will *not* be moved by the
script. This means that if an entry includes images, it's best to at
least partially complete the entry, run `mkblog.py` to move it to the
correct directory, then add the image files to the proper directory. Or,
you could just not be lazy and put the entry in the correct folder to
begin with...

The template now displays two links back to the index (one at the top,
one at the bottom) for all files except the index itself. The template
also includes the date the entry was posted for all but the index.