Fixed problem caused by last commit
In the last commit I removed the month name as one of the entry header
fields, but this was used to display the posted date on the actual entry
page. I was hoping to not have to type in the full name of the month
every time I write a new entry, so this works around needing to add it.