Newly-formatted blog entries
These changes to the Markdown blog entries allow Sphinx to capture the
metadata for each entry in a similar way that Caddy did with the JSON
headers.

Also, now that the blog appears in the site table of contents properly,
I've added year indexes, which are really just files that redirect to
the main blog index page but with the tag filter set to the selected
year.

Re #2