Blog year index pages now have table of contents
This avoids the warnings about the blog pages not being in a toctree.

One thing I didn't attempt was to remove the JavaScript that makes the
year link in the right hand menu `active` when viewing a blog post,
which should now be unnecessary now that Sphinx knows the structure.