Majority of CSE 469 materials in place
Still need to tweak some of the wording to the syllabus, other small
things, but the structure is in place and the aesthetics are what I had
hoped they would be.

Even if rtfd/sphinx_rtd_theme#699 never gets resolved, allowing
`navigation_depth` to not be limited to 2 (in `conf.py`) at least shows
the other pages available.