Removing two unnecessary fields from project spec
The Index field isn't used for anything. I think when I wrote the
project description I thought I would use it, but never got around to
verifying that it was actually useful. Removing it from the spec makes
it so no one has to try to figure out how to use it or even generate it
(the example index value was 0, which isn't a valid UUID... lots of
mixed messages there).

If I were to remove the Index field but leave the hash of the current
block, the blockchain file would end up storing the same hash twice, one
right after the other, which is pointless and a waste of space.